+++ seispy/plotting/utils.py
@@ -16,8 +16,9 @@ Python module with general utilities for making plotting easier.
 By: Shahar Shani-Kadmiel, August 2012, kadmiel@post.bgu.ac.il
 
 """
-
+from matplotlib import rc
 import numpy as np
+
 
 def pretty_ticks(min, max, number_of_ticks, show_zero=False):
     """Function for controlling tickmarks.
@@ -37,3 +38,36 @@ def pretty_ticks(min, max, number_of_ticks, show_zero=False):
         negative_ticks = np.arange(0,min,-step)[::-1]
         ticks = np.hstack((negative_ticks[:-1],positive_ticks))
     return ticks
+
+
+def set_matplotlib_rc_params():
+    """
+    Set matplotlib rcparams for plotting
+    """
+    font = {'family'        : 'sans-serif',
+            'sans-serif'    : 'Helvetica',
+            'style'         : 'normal',
+            'variant'       : 'normal',
+            'weight'        : 'medium',
+            'stretch'       : 'normal',
+            'size'          : 12.0}
+    rc('font', **font)
+
+    legend = {'fontsize'    : 10.0}
+    rc('legend', **legend)
+
+    axes = {'titlesize'     : 14.0,
+            'labelsize'     : 12.0}
+    rc('axes', **axes)
+    rc('pdf', fonttype=42)
+
+    ticks = {'direction'    : 'out',
+             'labelsize'    : 12.0,
+             'major.pad'    : 4,
+             'major.size'   : 5,
+             'major.width'  : 1.0,
+             'minor.pad'    : 4,
+             'minor.size'   : 2.5,
+             'minor.width'  : 0.75}
+    rc('xtick', **ticks)
+    rc('ytick', **ticks)
